refactor: update import paths to use utils/formatters
